Refactor move actions tests
