+++ tests/gui/test_widgets.py
@@ -3,10 +3,8 @@ Make sure the UI widgets are configured correctly and work as expected.
 """
 import math
 import random
-import unittest
 from datetime import datetime
 from gettext import gettext as _
-from unittest import mock
 from unittest.mock import Mock, PropertyMock
 
 import arrow
@@ -15,16 +13,11 @@ import sqlalchemy.orm.exc
 from PyQt5.QtCore import QEvent, QSize, Qt
 from PyQt5.QtGui import QFocusEvent, QMovie, QResizeEvent
 from PyQt5.QtTest import QTest
-from PyQt5.QtWidgets import QApplication, QMenu, QVBoxLayout, QWidget
+from PyQt5.QtWidgets import QApplication, QVBoxLayout, QWidget
 from sqlalchemy.orm import attributes, scoped_session, sessionmaker
 
-from securedrop_client import db, logic, state, storage
+from securedrop_client import db, logic, storage
 from securedrop_client.export import ExportError, ExportStatus
-from securedrop_client.gui.actions import (
-    DeleteConversationAction,
-    DeleteSourceAction,
-    DownloadConversation,
-)
 from securedrop_client.gui.source import DeleteSourceDialog
 from securedrop_client.gui.widgets import (
     ActivityStatusBar,
@@ -4738,72 +4731,6 @@ def test_ConversationView_add_not_downloaded_file(mocker, homedir, source, sessi
     assert isinstance(file_widget, FileWidget)
 
 
-def test_DeleteSourceAction_init(mocker):
-    mock_controller = mocker.MagicMock()
-    mock_source = mocker.MagicMock()
-    DeleteSourceAction(mock_source, None, mock_controller)
-
-
-def test_DeleteSourceAction_trigger(mocker):
-    mock_controller = mocker.MagicMock()
-    mock_source = mocker.MagicMock()
-    mock_delete_source_dialog_instance = mocker.MagicMock(DeleteSourceDialog)
-    mock_delete_source_dialog = mocker.MagicMock()
-    mock_delete_source_dialog.return_value = mock_delete_source_dialog_instance
-
-    mocker.patch("securedrop_client.gui.actions.DeleteSourceDialog", mock_delete_source_dialog)
-    delete_source_action = DeleteSourceAction(mock_source, None, mock_controller)
-    delete_source_action.trigger()
-    mock_delete_source_dialog_instance.exec.assert_called_once()
-
-
-def test_DeleteSourceAction_requires_an_authenticated_journalist(mocker):
-    mock_controller = mocker.MagicMock()
-    mock_controller.api = None  # no aouthenticated journalist
-    mock_source = mocker.MagicMock()
-    mock_delete_source_dialog_instance = mocker.MagicMock(DeleteSourceDialog)
-    mock_delete_source_dialog = mocker.MagicMock()
-    mock_delete_source_dialog.return_value = mock_delete_source_dialog_instance
-
-    mocker.patch("securedrop_client.gui.actions.DeleteSourceDialog", mock_delete_source_dialog)
-    delete_source_action = DeleteSourceAction(mock_source, None, mock_controller)
-    delete_source_action.trigger()
-    assert not mock_delete_source_dialog_instance.exec.called
-    mock_controller.on_action_requiring_login.assert_called_once()
-
-
-def test_DeleteConversationAction_trigger(mocker):
-    mock_controller = mocker.MagicMock()
-    mock_source = mocker.MagicMock()
-    mock_delete_conversation_dialog_instance = mocker.MagicMock(DeleteSourceDialog)
-    mock_delete_conversation_dialog = mocker.MagicMock()
-    mock_delete_conversation_dialog.return_value = mock_delete_conversation_dialog_instance
-
-    mocker.patch(
-        "securedrop_client.gui.actions.DeleteConversationDialog", mock_delete_conversation_dialog
-    )
-    delete_conversation_action = DeleteConversationAction(mock_source, None, mock_controller)
-    delete_conversation_action.trigger()
-    mock_delete_conversation_dialog_instance.exec.assert_called_once()
-
-
-def test_DeleteConversationAction_trigger_when_user_is_loggedout(mocker):
-    mock_controller = mocker.MagicMock()
-    mock_controller.api = None
-    mock_source = mocker.MagicMock()
-    mock_delete_conversation_dialog_instance = mocker.MagicMock(DeleteSourceDialog)
-    mock_delete_conversation_dialog = mocker.MagicMock()
-    mock_delete_conversation_dialog.return_value = mock_delete_conversation_dialog_instance
-
-    mocker.patch(
-        "securedrop_client.gui.conversation.DeleteConversationDialog",
-        mock_delete_conversation_dialog,
-    )
-    delete_conversation_action = DeleteConversationAction(mock_source, None, mock_controller)
-    delete_conversation_action.trigger()
-    mock_delete_conversation_dialog_instance.exec.assert_not_called()
-
-
 def test_DeleteSource_from_source_menu_when_user_is_loggedout(mocker):
     mock_controller = mocker.MagicMock()
     mock_controller.api = None
@@ -5747,113 +5674,3 @@ def test_SenderIcon_set_normal_styles_purple_for_authenticated_user(mocker):
     si.set_normal_styles()
 
     si.setObjectName.assert_called_once_with("SenderIcon_current_user")
-
-
-class TestDownloadConversation(unittest.TestCase):
-    def test_trigger(self):
-        menu = QMenu()
-        controller = mock.MagicMock()
-        app_state = state.State()
-        action = DownloadConversation(menu, controller, app_state)
-
-        conversation_id = state.ConversationId("some_conversation")
-        app_state.selected_conversation = conversation_id
-
-        action.trigger()
-
-        controller.download_conversation.assert_called_once_with(conversation_id)
-
-    def test_trigger_downloads_nothing_if_no_conversation_is_selected(self):
-        menu = QMenu()
-        controller = mock.MagicMock()
-        app_state = state.State()
-        action = DownloadConversation(menu, controller, app_state)
-
-        action.trigger()
-        assert controller.download_conversation.not_called
-
-    def test_gets_disabled_when_no_files_to_download_remain(self):
-        menu = QMenu()
-        controller = mock.MagicMock()
-        app_state = state.State()
-        action = DownloadConversation(menu, controller, app_state)
-
-        conversation_id = state.ConversationId(3)
-        app_state.selected_conversation = conversation_id
-
-        app_state.add_file(conversation_id, 5)
-        app_state.file(5).is_downloaded = True
-
-        action.setEnabled(True)  # only for extra contrast
-        app_state.selected_conversation_files_changed.emit()
-        assert not action.isEnabled()
-
-    def test_gets_enabled_when_files_are_available_to_download(self):
-        menu = QMenu()
-        controller = mock.MagicMock()
-        app_state = state.State()
-        action = DownloadConversation(menu, controller, app_state)
-
-        conversation_id = state.ConversationId(3)
-        app_state.selected_conversation = conversation_id
-
-        app_state.add_file(conversation_id, 5)
-        app_state.file(5).is_downloaded = False
-
-        action.setEnabled(False)  # only for extra contrast
-        app_state.selected_conversation_files_changed.emit()
-        assert action.isEnabled()
-
-    def test_gets_initially_disabled_when_file_information_is_available(self):
-        menu = QMenu()
-        controller = mock.MagicMock()
-        app_state = state.State()
-
-        conversation_id = state.ConversationId(3)
-        app_state.selected_conversation = conversation_id
-        app_state.add_file(conversation_id, 5)
-        app_state.file(5).is_downloaded = True
-
-        action = DownloadConversation(menu, controller, app_state)
-
-        assert not action.isEnabled()
-
-    def test_gets_initially_enabled_when_file_information_is_available(self):
-        menu = QMenu()
-        controller = mock.MagicMock()
-        app_state = state.State()
-
-        conversation_id = state.ConversationId(3)
-        app_state.selected_conversation = conversation_id
-        app_state.add_file(conversation_id, 5)
-        app_state.file(5).is_downloaded = False
-
-        action = DownloadConversation(menu, controller, app_state)
-
-        assert action.isEnabled()
-
-    def test_does_not_require_state_to_be_defined(self):
-        menu = QMenu()
-        controller = mock.MagicMock()
-        action = DownloadConversation(menu, controller, app_state=None)
-
-        action.setEnabled(False)
-        assert not action.isEnabled()
-
-        action.setEnabled(True)
-        assert action.isEnabled()
-
-    def test_on_selected_conversation_files_changed_handles_missing_state_gracefully(
-        self,
-    ):
-        menu = QMenu()
-        controller = mock.MagicMock()
-        action = DownloadConversation(menu, controller, None)
-
-        action.setEnabled(True)
-        action._on_selected_conversation_files_changed()
-        assert action.isEnabled()
-
-        action.setEnabled(False)
-        action._on_selected_conversation_files_changed()
-        assert not action.isEnabled()
